Make `from zschema import *` great again.
Do not merge yet.

This fixes the issue where `from zschema import *` was not importing all
relevant names (e.g. SubRecord). However, zschema validate is currently
not working because the zschema.registry module seems to be loaded
twice, and then the lookup for the relevent schema in __schemas
NameError's. I do not understand why this is happening, or how to fix.
Even reverting to the old code in __init__.py results in the same
double-initialize behavior.