Add zschema.registry.all_schemas
This function returns all schemas defined in the registry. It returns a
shallow copy of the internal __zschema_schemas object, so that adding
items to the returned dict does not modify the internal registry.

This function can be used to genenerate all the schemas in a file
programmatically, and may be useful in CI scripts.