Enums without possible values have an empty values property
Before this commit, an Enum leaf without values (used in a schema like `Enum()`) would have `values` and `values_s` properties that were a list/set containing a single value, `None`. This didn’t make sense to me, and it complicated producing clean documentation output.

Instead, for the “any values are accepted” case, we’ll keep those two Enum properties as an empty list & set, and check for this specific case when validating a document against a schema.

Finally, this PR cleans up the code that produces documentation for Enum leaves.