Add Python3 compatibility
This updates all code to be Python 3 compatible. As a side effect, it
removes support for generating Protobufs with implicit field numbers.
This feature was dangerous, and as far as I know, unused.

The following changes add Python3 compatiblity:
- Use future.print_function
- Replace all '%' string interpolation with .format()
- Use isinstance() instead of directly comparing types
- Compare against six.string_type, instead of (str, unicode)
- Replace .iteritems() with .items()
- Import Python3 dictionaries, ints, and strings from builtins
- Replace longs with builtin.int
- Define __bool__ for everything that had __nonzero__
- Define __lt__ and __gt__ for everything that had __cmp__
- Refactor code to avoid sorting lists or dicts where keys are None
- Update assertRaises calls in tests to correctly specify which
  exception should be thrown

Note: There is no performance hit from using .items() when builtins.dict
is imported.