SubRecordType()() objects behave like objects
Every object created by a type returned by SubRecordType effectively
stored all of it's attributes as class attributes, which meant that if
you did the following:

```
MyType = SubRecordType()
a = MyType()
b = MyType()
```

...then changes to `a` would be reflected on `b`, which broke isolation
of types in schemas.

This reset properties defined in the call to SubRecordType() at object
initialization and applies them to the instances, which should prevent
weird interactions between objects.